feat: 引入 IDE 命令工具类，优化命令测试和执行功能 [Feature|IDE] #5
